Move appeal domain event logic into new service
Moved the logic into a standalone service for consistency with other domain events.

This change has been made in preparation for adding changes to the domain event.

I also added some additional checks to the AppealsService tests to ensure only a success or failure email is sent
