+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/service/AppealService.kt
@@ -1,15 +1,8 @@
 package uk.gov.justice.digital.hmpps.approvedpremisesapi.service
 
-import org.springframework.beans.factory.annotation.Value
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.AssessmentAppealed
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.AssessmentAppealedEnvelope
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.PersonReference
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.StaffMember
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.AppealDecision
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.ClientResult
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.CommunityApiClient
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.AppealEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.AppealRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ApplicationEntity
@@ -17,26 +10,21 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ApprovedPremi
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.AssessmentEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserEntity
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserRole
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.DomainEvent
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.validated
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.AuthorisableActionResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.ValidatableActionResult
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Cas1AppealDomainEventService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Cas1AppealEmailService
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.UrlTemplate
 import java.time.LocalDate
 import java.time.OffsetDateTime
 import java.util.UUID
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.AppealDecision as DomainEventApiAppealDecision
 
 @Service
 class AppealService(
   private val appealRepository: AppealRepository,
   private val assessmentService: AssessmentService,
-  private val domainEventService: DomainEventService,
-  private val communityApiClient: CommunityApiClient,
   private val cas1AppealEmailService: Cas1AppealEmailService,
-  @Value("\${url-templates.frontend.application}") private val applicationUrlTemplate: UrlTemplate,
-  @Value("\${url-templates.frontend.application-appeal}") private val applicationAppealUrlTemplate: UrlTemplate,
+  private val cas1AppealDomainEventService: Cas1AppealDomainEventService,
 ) {
   fun getAppeal(appealId: UUID, application: ApplicationEntity): AuthorisableActionResult<AppealEntity> {
     val appeal = appealRepository.findByIdOrNull(appealId)
@@ -90,7 +78,7 @@ class AppealService(
           ),
         )
 
-        saveEvent(appeal)
+        cas1AppealDomainEventService.appealRecordCreated(appeal)
 
         when (decision) {
           AppealDecision.accepted -> {
@@ -106,58 +94,4 @@ class AppealService(
       },
     )
   }
-
-  private fun saveEvent(appeal: AppealEntity) {
-    val id = UUID.randomUUID()
-    val timestamp = appeal.createdAt.toInstant()
-
-    val staffDetails = when (val result = communityApiClient.getStaffUserDetails(appeal.createdBy.deliusUsername)) {
-      is ClientResult.Success -> result.body
-      is ClientResult.Failure -> result.throwException()
-    }
-
-    domainEventService.saveAssessmentAppealedEvent(
-      DomainEvent(
-        id = id,
-        applicationId = appeal.application.id,
-        assessmentId = null,
-        bookingId = null,
-        crn = appeal.application.crn,
-        occurredAt = timestamp,
-        data = AssessmentAppealedEnvelope(
-          id = id,
-          timestamp = timestamp,
-          eventType = "approved-premises.assessment.appealed",
-          eventDetails = AssessmentAppealed(
-            applicationId = appeal.application.id,
-            applicationUrl = applicationUrlTemplate.resolve("id", appeal.application.id.toString()),
-            appealId = appeal.id,
-            appealUrl = applicationAppealUrlTemplate.resolve(
-              mapOf("applicationId" to appeal.application.id.toString(), "appealId" to appeal.id.toString()),
-            ),
-            personReference = PersonReference(
-              crn = appeal.application.crn,
-              noms = appeal.application.nomsNumber!!,
-            ),
-            deliusEventNumber = (appeal.application as ApprovedPremisesApplicationEntity).eventNumber,
-            createdAt = timestamp,
-            createdBy = StaffMember(
-              staffCode = staffDetails.staffCode,
-              staffIdentifier = staffDetails.staffIdentifier,
-              forenames = staffDetails.staff.forenames,
-              surname = staffDetails.staff.surname,
-              username = staffDetails.username,
-            ),
-            appealDetail = appeal.appealDetail,
-            decision = parseDecision(appeal.decision),
-            decisionDetail = appeal.decisionDetail,
-          ),
-        ),
-      ),
-    )
-  }
-
-  private fun parseDecision(value: String): DomainEventApiAppealDecision =
-    DomainEventApiAppealDecision.entries.firstOrNull { it.value == value }
-      ?: throw IllegalArgumentException("Unknown appeal decision type '$value'")
 }
